project/requirements.txt
pyyaml
structlog
openai==0.28
pytest